requirements.txt
boto3==1.4.4
botocore==1.5.68
certifi==2017.4.17
chardet==3.0.4
dj-database-url==0.4.2
dj-static==0.0.6
Django==1.11.2
django-toolbelt==0.0.1
djangorestframework==3.6.3
docutils==0.13.1
gunicorn==19.7.1
idna==2.5
jmespath==0.9.3
psycopg2==2.7.1
python-dateutil==2.6.0
pytz==2017.2
requests==2.18.1
s3transfer==0.1.10
six==1.10.0
static3==0.7.0
urllib3==1.21.1
whitenoise==3.3.0
